Added subtasks and added more data validation

+++ server/src/task/task.controller.ts
@@ -6,6 +6,8 @@ import {
   Patch,
   Param,
   Delete,
+  HttpStatus,
+  HttpException
 } from '@nestjs/common';
 import { TaskService } from './task.service';
 import { CreateTaskDto } from './dto/create-task.dto';
@@ -17,6 +19,18 @@ export class TaskController {
 
   @Post()
   create(@Body() createTaskDto: CreateTaskDto) {
+    for(let i = 0; i < createTaskDto.dependencies.length; i++) {
+      if (createTaskDto.dependencies[i].length !== 36) {
+        throw new HttpException('Invalid dependency ID format!', HttpStatus.BAD_REQUEST);
+      }
+    }
+
+    for(let i = 0; i < createTaskDto.subtasks.length; i++) {
+      if (createTaskDto.subtasks[i].length !== 36) {
+        throw new HttpException('Invalid subtask ID format!', HttpStatus.BAD_REQUEST);
+      }
+    }
+
     return this.taskService.create(createTaskDto);
   }
 
@@ -27,21 +41,49 @@ export class TaskController {
 
   @Get('/author/:id')
   findByAuthor(@Param('id') id: string) {
-    return this.taskService.findByAuthor(id);
+    if (id.length === 36) {
+      return this.taskService.findByAuthor(id);
+    } else {
+      throw new HttpException('Invalid ID format!', HttpStatus.BAD_REQUEST);
+    }
   }
 
   @Get(':id')
   findOne(@Param('id') id: string) {
-    return this.taskService.findOne(id);
+    if (id.length === 36) {
+      return this.taskService.findOne(id);
+    } else {
+      throw new HttpException('Invalid ID format!', HttpStatus.BAD_REQUEST);
+    }
   }
 
   @Patch(':id')
   update(@Param('id') id: string, @Body() updateTaskDto: UpdateTaskDto) {
-    return this.taskService.update(id, updateTaskDto);
+    for(let i = 0; i < updateTaskDto.dependencies.length; i++) {
+      if (updateTaskDto.dependencies[i].length !== 36) {
+        throw new HttpException('Invalid dependency ID format!', HttpStatus.BAD_REQUEST);
+      }
+    }
+
+    for(let i = 0; i < updateTaskDto.subtasks.length; i++) {
+      if (updateTaskDto.subtasks[i].length !== 36) {
+        throw new HttpException('Invalid subtask ID format!', HttpStatus.BAD_REQUEST);
+      }
+    }
+
+    if (id.length === 36) {
+      return this.taskService.update(id, updateTaskDto);
+    } else {
+      throw new HttpException('Invalid ID format!', HttpStatus.BAD_REQUEST);
+    }
   }
 
   @Delete(':id')
   remove(@Param('id') id: string) {
-    return this.taskService.remove(id);
+    if (id.length === 36) {
+      return this.taskService.remove(id);
+    } else {
+      throw new HttpException('Invalid ID format!', HttpStatus.BAD_REQUEST);
+    }
   }
 }
